altered show.py - might be controversial
This commit cleans up the imports. It also cleans up the inital masked
array used by show, to be easier to read. It no longer requires
mask=mask==False in areas hidden to the user.

This commit also overloads the figax parameter. Previously it took in
None or a figure and axis handle. Now it can also take in an integer,
which describes a specific plot number. Then this plot number can be
replotted over, so that for example, you could always plot on figure(3).

The commit also changes the plotting of images, in a way that _should_
not break things. First, ax.matshow has been changed to ax.imshow, which
is the more universal function. Then, the mask handling has changed.
Looking at the code is the most helpful thing to do, but essentially, if
a mask is found, and the color is not None, then the colormap is copied,
the color for 'bad' values is updated to the mask color, and that is
all. Outside of this if statement the imshow command is run. This allows
you to 1) set None as a color so you can plot masked arrays on top of
each other (important for show_amorphous_ring_fit) 2) not have to worry
about other masks when redrawing multiple figures on top of each other.