modified show_image_grid, added show_nn
added the inputs title and num to show_image_grid. The first allows you
to iterate through a list of titles, the second lets you choose a figure
number to plot on. I also simplified the iteration through the axes, and
removed plt.tight_layout if you return the figure and axes, as that
often causes problems if not called last.

I also added show_nn, which is a very simple function to just plot a 3x3
grid of probe positions with a common mask. This is very nice to quickly
change contrast, or make composite images that are just 1 image.