modified show_amorphous_ring
This commit changes show_amorphous_ring, fixing some bugs and improving
some functionality. First, default values are added for qmin and qmax,
such that when unchanged, it plots wedges extending from the center to
the very edge of the image. Three more inputs were added, maskcenter,
mask and fignum. When maskcenter is True, values < qmin are set to the
average value > qmin. This is helpful for when you want to use normal
scaling and have a bright center spot. mask is a mask that replaces data
with the minimum values outside of the mask, helpful when you have
diffraction spots to mask off in the pattern. Finally fignum allows you
to replot consistently on the same matplotlib figure, overwriting it as
you go.