This commit changes clipval computation
This commit changes how the clipvals are computed when a mask is
passed into py4DSTEM.show. Previously it did not take into account
masked areas, so that if you masked off the central beam, its
intensity was still being used for computing vmin and vmax. So in
the situation where you called `py4DSTEM.show(ar, mask=mask)`, you
would see nothing since the vmax was still scaling to the central
beam's intensity. Now the mask that is passed in is combined with
the mask from the scaling operation before clipvals are computed,
which is a better default.