Show fix to compute clipvals after masking (#343)
* This commit changes clipval computation

This commit changes how the clipvals are computed when a mask is
passed into py4DSTEM.show. Previously it did not take into account
masked areas, so that if you masked off the central beam, its
intensity was still being used for computing vmin and vmax. So in
the situation where you called `py4DSTEM.show(ar, mask=mask)`, you
would see nothing since the vmax was still scaling to the central
beam's intensity. Now the mask that is passed in is combined with
the mask from the scaling operation before clipvals are computed,
which is a better default.

* This commit fixes the show.ipynb for v0.13.6

This commit fixes function definition changes in the `show.ipynb`
file to match v0.13.6 and changes the function definition in
`show_ellipses` to match what `add_ellipses` expects. This is only
changing the ellipse equation to accept (a,b,theta) instead of
(a,e,theta).

* tiny bugfix because you want the array when changing the mask to a boolean, not the py4DSTEM Array object

Co-authored-by: SE Zeltmann <steven.zeltmann@lbl.gov>