Fixes plotting drawn elements blocking others
This commit fixes the show_geometry functions. Previously, there
was a plt.show() call if the `returnfig` argument was `True`. This
was a mistake, as that blocks future draws to the same figure.

By removing the plt.show call, the same figure can be drawn on
multiple times, which is helpful when working with subplots.

I have tested this in both a Jupyter notebook as well as the
IPython command line to work effectively. If you don't pass in
a figax tuple, new figures are made as expected.

This should probably be tested by someone else before being merged.


Former-commit-id: ba9ced47eedd180ceaf7d35b7c2ac24a260bfa10